requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
drf-spectacular==0.28.0
oracledb==3.1.1
gunicorn>=21.2
